edited the site as requested

+++ mainsite/models.py
@@ -13,7 +13,7 @@ class Winners(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE)
     number = models.FloatField(name="number")
     class_number = models.CharField(name='class_number', max_length=10)
-    delta = models.PositiveSmallIntegerField()
+    delta = models.FloatField()
 
 
 class Number(models.Model):
